move DateNodeRouter to root package
